Create religious-primes-and-morality
This was my very first research study looking at the priming effects of different religious cognitions on endorsements of various moral domains. The whole study was a mess with regards to data collection and so that's why you see a long- and short-version due to some variables missing from wave to wave. This was done through R/RStudio and it's pretty clear that this was done when I was starting to teach myself R.

Later on, I re-visited this particular study and incorporated some Bayesian approaches to the regression model using the rstan package.